repo-updater: Factor out Sources type
